change location and name of plot results
